Mise en place API REST PLUS TEST SUR POSTMAN

+++ src/main/java/com/ruffin/Spring_Mvc_Restful_GraphQl/entity/Release.java
@@ -13,7 +13,7 @@ public class Release {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Integer id;
+	private Long id;
 	
 	private String releaseDate;
 	private String description;
@@ -22,17 +22,17 @@ public class Release {
 		super();
 	}
 	
-	public Release(Integer id, String releaseDate, String description) {
+	public Release(Long id, String releaseDate, String description) {
 		super();
 		this.id = id;
 		this.releaseDate = releaseDate;
 		this.description = description;
 	}
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getReleaseDate() {
